git corrections for rl2_cutline,rl2_berlin and common

+++ rl2_cutline/create_db/rl2_create_eiskeller_db.sh
@@ -3,28 +3,58 @@
 #---------------------------------------------------
 # Adapt db/sql diretory path where needed
 #---------------------------------------------------
+IMAGE_DIR="../berlin_images";
 DB_DIR="../source_db";
 SQL_DIR="../source_sql";
 #---------------------------------------------------
 # Adapt db/sql file name where needed
 #---------------------------------------------------
-DB_FILE="1926.berlin_eiskeller.db";
-SQL_FILE="create.rl2.1926.berlin_eiskeller.sql";
+IMAGE_BASE="1926.berlin_eiskeller";
 #---------------------------------------------------
 # no more changes should be needed
 #---------------------------------------------------
+DB_FILE="${IMAGE_BASE}.db";
+SQL_FILE="create.rl2.${IMAGE_BASE}.sql";
+IMAGE_FILE="${IMAGE_BASE}.4000.3068.tif";
+# IMAGE_FILE="${IMAGE_FILE}.4000.3068.info.txt";
 CMD_SQLITE3=`which sqlite3`;
 export "SPATIALITE_SECURITY=relaxed"
 DB_PATH="${DB_DIR}/${DB_FILE}";
 SQL_PATH="${SQL_DIR}/${SQL_FILE}";
+IMAGE_PATH="${IMAGE_DIR}/${IMAGE_FILE}";
 #---------------------------------------------------
-if [ -f "${DB_PATH}" ]
+if [ ! -f "${IMAGE_PATH}" ]
 then
- rm ${DB_PATH}
+ exit_rc=1;
+ #---------------------------------------------------
+ WWW_BERLIN_IMAGES="http://www.mj10777.de/rasterlite2/build_berlin_data/berlin_images";
+ WWW_PATH="${WWW_BERLIN_IMAGES}/${IMAGE_FILE}";
+ CMD_WGET=`which wget`;
+ if [ ! -z "${CMD_WGET}" ]
+ then 
+  echo "-I-> downloading needed 240 MB Image-File[${IMAGE_PATH}]";
+  ${CMD_WGET} ${WWW_PATH} -O ${IMAGE_PATH}
+  exit_rc=$?
+  echo "${CMD_WGET} rc=$exit_rc ";
+  if [ -f "${IMAGE_PATH}" ]
+  then
+   exit_rc=0;
+  fi
+ fi
+ #---------------------------------------------------
 fi
 #---------------------------------------------------
-${CMD_SQLITE3} ${DB_PATH} < ${SQL_PATH}
-exit_rc=$?
+if [ "$exit_rc" -eq "0" ]
+then
+ if [ -f "${DB_PATH}" ]
+ then
+  rm ${DB_PATH}
+ fi
+ #---------------------------------------------------
+ ${CMD_SQLITE3} ${DB_PATH} < ${SQL_PATH}
+ exit_rc=$?
+ #---------------------------------------------------
+fi
 #---------------------------------------------------
 BENE="'Tutto bene!'";
 BENENO="'No bene!'";
